Got rid of nested fsm, and moved fsm to stp.utils
